+++ fs/rsext4/src/entries/mod.rs
@@ -3,9 +3,9 @@
 pub mod classic_dir;
 mod dir_entry;
 mod disk;
-pub mod htree_dir;
+pub(crate) mod htree_dir;
 mod iterator;
 
 pub use dir_entry::{Ext4DirEntry, Ext4DirEntry2, Ext4DirEntryTail, Ext4ExtentStatus};
-pub use htree_dir::{Ext4DxCountlimit, Ext4DxEntry, Ext4DxNode, Ext4DxRoot, Ext4DxRootInfo};
+pub(crate) use htree_dir::{Ext4DxEntry, Ext4DxRootInfo};
 pub use iterator::{DirEntryIterator, Ext4DirEntryInfo};