+++ fs/rsext4/src/lib.rs
@@ -74,7 +74,7 @@ pub mod error;
 pub mod ext4;
 pub mod extents_tree;
 pub mod file;
-pub mod hashtree;
+pub(crate) mod hashtree;
 mod indirect;
 pub mod io;
 pub mod jbd2;